+++ CHANGELOG.md
@@ -16,6 +16,10 @@ Documentation for TransferBench is available at
   0 will run over a range of CUs to use, and report only the results of the best one found. This can be tuned
   for performance by setting the MIN_VAR_SUBEXEC and MAX_VAR_SUBEXEC environment variables to narrow the
   search space.  The number of CUs used will be identical for all variable subExecutor transfers
+- Experimental new "healthcheck" preset config which currently only supports MI300 series.  This preset runs
+  through CPU to GPU bandwidth tests and all-to-all XGMI bandwidth tests and compares against expected values
+  Pass criteria limits can be modified (due to platform differences) via the environment variables
+  LIMIT_UDIR (undirectional), LIMIT_BDIR (bidirectional), and LIMIT_A2A (Per GPU-GPU link bandwidth)
 
 ### Fixed
 - Fixed out-of-bounds memory access during topology detection that can happen if the number of